u64->i32

+++ src/main.rs
@@ -146,12 +146,12 @@ fn read_local_file_into_memory(input_file: &PathBuf) ->Result<Cursor<Vec<u8>>, E
 
 
 
-fn count_from_compressed_data(compressed_data: Vec<u8>, mut local_counter: HashMap<u64, usize>) -> Result<HashMap<u64, usize>, Error> {
+fn count_from_compressed_data(compressed_data: Vec<u8>, mut local_counter: HashMap<i32, usize>) -> Result<HashMap<i32, usize>, Error> {
     let mut decoder = MultiGzDecoder::new(&compressed_data[..]);
     let mut decompressed_data = Vec::new();
     decoder.read_to_end(&mut decompressed_data).unwrap();
 
-    let numbers: Vec<u64> = from_slice(&decompressed_data).unwrap();
+    let numbers: Vec<i32> = from_slice(&decompressed_data).unwrap();
     for number in numbers {
         *local_counter.entry(number).or_insert(0) += 1;
     }
@@ -160,7 +160,7 @@ fn count_from_compressed_data(compressed_data: Vec<u8>, mut local_counter: HashM
 
 
 
-async fn process_file(input: &PathBuf, token_counter: &Arc<DashMap<u64, usize>>) -> Result<(), Error> {
+async fn process_file(input: &PathBuf, token_counter: &Arc<DashMap<i32, usize>>) -> Result<(), Error> {
     // Count number of tokens in each context in each input 
 
 
@@ -170,7 +170,7 @@ async fn process_file(input: &PathBuf, token_counter: &Arc<DashMap<u64, usize>>)
         BufReader::new(read_local_file_into_memory(&input).unwrap())
     };
     let mut tar = Archive::new(reader);
-    let mut local_token_counter : HashMap<u64, usize> = HashMap::new(); 
+    let mut local_token_counter : HashMap<i32, usize> = HashMap::new(); 
 
 
     for entry in tar.entries()? {
@@ -249,7 +249,7 @@ fn main() -> Result<()> {
 
     // Step 3: finalize the dashmap into something we can save
     //token_counter.into_inner();
-    let token_counter : HashMap<u64, usize> = <DashMap<u64, usize> as Clone>::clone(&token_counter).into_iter().collect();
+    let token_counter : HashMap<i32, usize> = <DashMap<i32, usize> as Clone>::clone(&token_counter).into_iter().collect();
     let json_data = json!(token_counter);
